study_buddy/pyproject.toml
[project]
name = "study-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.5.9",
    "faiss-cpu>=1.11.0",
    "numpy>=2.2.6",
    "ollama>=0.5.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "starlette>=0.47.0",
    "streamlit>=1.45.1",
    "typer>=0.16.0",
    "uvicorn>=0.34.3",
]
